% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_impute.r
\name{average_impute}
\alias{average_impute}
\title{Get Averages of MICE Imputed Values}
\usage{
average_impute(data, id_column, imputed_variables, predictors, ...)
}
\arguments{
\item{data}{The tibble containing data to be imputed.}

\item{id_column}{The index column of the data tibble.}

\item{imputed_variables}{The variables to impute and the imputation method, as a named list---e.g. list('weight' = 'pmm', 'height' = 'pmm'). Methods are form MICE’s \code{mice()} function}

\item{predictors}{The predictors to use to impute variables.}

\item{...}{Additional paramters passed to MICE’s \code{mice()} function, e.g. \code{m} and \code{seed}. See the documentation for \code{mice()}.}
}
\value{
A tibble with and index column and descriptive statistics for each imputed column.
}
\description{
Runs MICE imputation using specified predictors over set of tibble columns and returns descriptive statistics over the run.
}
\details{
For example if the column weight is imputed, the function will return min_weight, mean_weight, median_weight, and max_weight calculated from all the imputations performed, so if MICE is set to run 200 times, the value returned will be an average of all those runs.
}
\examples{
tibble::tibble(
  a=c(1,2,3),
  b=c(1,NA,3),
  c=c(1,2,3)
) |>
  average_impute(
    'a',
    list('b' = 'pmm'),
    c('a', 'c')
   )
}
\seealso{
\code{\link[mice:mice]{mice::mice()}}
}
