% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_and_format.r
\name{count_and_format}
\alias{count_and_format}
\title{Count and Format}
\usage{
count_and_format(input_tibble)
}
\arguments{
\item{input_tibble}{A tibble.}
}
\value{
A string.
}
\description{
Takes a tibble, counts its rows, and formats the value with thousand dividers.
}
\examples{
data.frame(a = c(1,2,3)) |> count_and_format()
}
