% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr_capa.R
\name{egfr_capa}
\alias{egfr_capa}
\title{Calculate Cystatin C Based eGFR Using the CAPA Equation}
\usage{
egfr_capa(cystatin_c, age)
}
\arguments{
\item{cystatin_c}{Plasma cystatin C in mg/l.}

\item{age}{Age in years.}
}
\value{
eGFR in ml/min/1.73 m\eqn{^2}
}
\description{
Calculate estimated glomerular filtration rate (eGFR) using the Caucasian, Asian, Pediatric, and Adult (CAPA) formula described by Grubb et al\eqn{^1}.
}
\details{
\eqn{\text{eGFR} = 130 \times \text{cystatin C}^{-1.069} \times \text{age}^{-0,117} - 7}
}
\examples{
egfr_capa(.92, 25)
}
\references{
\enumerate{
\item Grubb A, Horio M, Hansson LO, Björk J, Nyman U, Flodin M, et al. Generation of a new cystatin C-based estimating equation for glomerular filtration reate by use of 7 assays standardized to the international calibrator. Clinical Chemistry. 2014 Jul 1;60(7):974–86.
}
}
