% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr_lm_revised.R
\name{egfr_lm_revised}
\alias{egfr_lm_revised}
\title{Calculate Creatinine Based eGFR Using the LM Revised Equation}
\usage{
egfr_lm_revised(creatinine, age, female)
}
\arguments{
\item{creatinine}{Plasma creatinine in µmol/l.}

\item{age}{Age in years.}

\item{female}{Sex. \code{TRUE} if female, \code{FALSE} if male.}
}
\value{
eGFR in ml/min/1.73 m\eqn{^2}
}
\description{
Calculate estimated glomerular filtration rate (eGFR) using the Revised Lund-Malmö equation without body weight measure (LM Revised) described by Björk et al\eqn{^1}.
}
\details{
\eqn{\text{eGFR} = e^{X - 0.0158 \times \text{age} + 0.438 \times \ln (\text{age})}}\tabular{ccc}{
   \strong{Sex} \tab \strong{Plasma Creatinine (µmol/l)} \tab \strong{X} \cr
   Female \tab <150 \tab \eqn{X = 2.50 + 0.0121 \times (150 - \text{creatinine})} \cr
   Female \tab ≥150 \tab \eqn{X = 2.50 - 0.926 \times \ln (\frac{\text{creatinine}}{150})} \cr
   Male \tab <180 \tab \eqn{X = 2.56 + 0.00968 \times (180 - \text{creatinine})} \cr
   Male \tab ≥180 \tab \eqn{X = 2.56 - 0.926 \times \ln (\frac{\text{creatinine}}{180})} \cr
}
}
\references{
\enumerate{
\item Björk J, Grubb A, Sterner G, Nyman U. Revised equations for estimating glomerular filtration rate based on the Lund-Malmö Study cohort. Scandinavian Journal of Clinical and Laboratory Investigation. 2011 May;71(3):232–9.
}
}
