% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/extract_terms_from_backwards_elimination_table.R
\name{extract_terms_from_backwards_elimination_table}
\alias{extract_terms_from_backwards_elimination_table}
\title{Extract Terms from Backwards Elimination Table}
\usage{
extract_terms_from_backwards_elimination_table(
  backwards_elimination_table,
  independent_variables_column_name = "independent_variables"
)
}
\arguments{
\item{backwards_elimination_table}{Tibble returned from \code{make_backwards_elimination_table()}.}

\item{independent_variables_column_name}{The name of the column with the terms. Defaults to 'independent_variables'.}
}
\value{
A list of names of terms.
}
\description{
Returns the remaining terms from a table returned from the function \code{make_backwards_elimination_table()}.
}
\examples{
tibble::tibble(
  independent_variables = c('(Intercept)', 'abc', 'def'),
  p_values = c(.5, .5, .5)
) |>
extract_terms_from_backwards_elimination_table()
}
