% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kva_codes_dialysis_catheter.R
\docType{data}
\name{kva_codes_dialysis_catheter}
\alias{kva_codes_dialysis_catheter}
\title{Swedish Procedure (KVÅ) Codes Regarding Hemodialysis Catheters}
\format{
An object of class \code{character} of length 19.
}
\source{
\url{https://www.socialstyrelsen.se/globalassets/sharepoint-dokument/dokument-webb/klassifikationer-och-koder/kva-inkl-beskrivningstexter.xlsx}
}
\usage{
kva_codes_dialysis_catheter
}
\description{
A list of \href{https://www.socialstyrelsen.se/statistik-och-data/klassifikationer-och-koder/kva/}{Swedish Procedure (KVÅ) Codes} indicating placement of a hemodialysis catheter.
}
\details{
The codes included in the list represent the following procedures.\tabular{ll}{
   \strong{Code} \tab \strong{Description} \cr
   DP002 \tab Inläggande av dialyskateter i vena femoralis \cr
   SP311 \tab Inläggning av central venkateter för dialys, i vena jugularis interna, icke tunnelerad \cr
   SP312 \tab Inläggning av central venkateter för dialys, i vena jugularis interna, tunnelerad \cr
   SP321 \tab Inläggning av central venkateter för dialys, i vena jugularis externa, icke tunnelerad \cr
   SP322 \tab Inläggning av central venkateter för dialys, i vena jugularis externa, tunnelerad \cr
   SP331 \tab Inläggning av central venkateter för dialys, i vena subclavia, icke tunnelerad \cr
   SP332 \tab Inläggning av central venkateter för dialys, i vena subclavia, tunnelerad \cr
   SP341 \tab Inläggning av central venkateter för dialys, i vena brachiocephalica, icke tunnelerad \cr
   SP342 \tab Inläggning av central venkateter för dialys, i vena brachiocephalica, tunnelerad \cr
   SP351 \tab Inläggning av central venkateter för dialys, i vena femoralis, icke tunnelerad \cr
   SP352 \tab Inläggning av central venkateter för dialys, i vena femoralis, tunnelerad \cr
   SP361 \tab Inläggning av central venkateter för dialys, i vena umbilicalis, icke tunnelerad \cr
   SP362 \tab Inläggning av central venkateter för dialys, i vena umbilicalis, tunnelerad \cr
   SP371 \tab Byte av central venkateter för dialys, icke tunnelerad \cr
   SP372 \tab Byte av central venkateter för dialys, tunnelerad \cr
   SP381 \tab Avlägsnande av central venkateter för dialys, icke tunnelerad \cr
   SP382 \tab Avlägsnande av central venkateter för dialys, tunnelerad \cr
   SP391 \tab Inläggning av central venkateter för dialys, i annat eller kärl UNS, icke tunnelerad \cr
   SP392 \tab Inläggning av central venkateter för dialys, i annat eller kärl UNS, tunnelerad \cr
}
}
\keyword{datasets}
