% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kva_codes_dialysis_treatment.R
\docType{data}
\name{kva_codes_dialysis_treatment}
\alias{kva_codes_dialysis_treatment}
\title{Swedish Procedure (KVÅ) Codes Regarding Hemodialysis Treatment}
\format{
An object of class \code{character} of length 13.
}
\source{
\url{https://www.socialstyrelsen.se/globalassets/sharepoint-dokument/dokument-webb/klassifikationer-och-koder/kva-inkl-beskrivningstexter.xlsx}
}
\usage{
kva_codes_dialysis_treatment
}
\description{
A list of \href{https://www.socialstyrelsen.se/statistik-och-data/klassifikationer-och-koder/kva/}{Swedish Procedure (KVÅ) Codes} indicating hemodialysis treatment.
}
\details{
The codes included in the list represent the following procedures.\tabular{ll}{
   \strong{Code} \tab \strong{Description} \cr
   FXP00 \tab Hemofiltration eller hemodialys under kardiopulmonal bypass \cr
   DR015 \tab Hemodialys, akut \cr
   DR016 \tab Hemodialys, kronisk \cr
   DR055 \tab Citratdialys \cr
   DR056 \tab Heparinfri dialys \cr
   DR060 \tab Hemhemodialyskontroll \cr
   DR061 \tab Hemhemodialysstart \cr
}
}
\keyword{datasets}
