% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_backwards_elimination_table.R
\name{make_backwards_elimination_table}
\alias{make_backwards_elimination_table}
\title{Perform Backward Elimination of Terms in a Linear Regression Model}
\usage{
make_backwards_elimination_table(
  tibble_to_analyze,
  dependent_variable,
  independent_variables,
  p_value_limit = 0.05,
  p_value_column_prefix = "p_values_run_"
)
}
\arguments{
\item{tibble_to_analyze}{The tibble to analyze.}

\item{dependent_variable}{The name of the column containing the dependent variable.}

\item{independent_variables}{A list of names of columns containing the independent variables.}

\item{p_value_limit}{The p-value limit.}

\item{p_value_column_prefix}{The prefix for the p-value columns in the output. Suffixed by run number.}
}
\value{
A tibble of runs, p-values, and remaining terms.
}
\description{
Takes a tibble of independent and dependent variables, makes a linear model, eliminates the term with the highest p-value, makes a new linear model and so on, until the term with the highest p-value is under a p-value limit. Outputs a tibble documenting the runs, p-values, and remaining terms.
}
\examples{
tibble::tibble(
  x1 = rnorm(50, 5, 2),
  x2 = rnorm(50, 3, 1.5),
  x3 = rnorm(50, 10, 5),
  x4 = rnorm(50, 0, 1),
  y = 3 + 2 * x1 - 1 * x2 + 0.5 * x3 + rnorm(50, 0, 1)
) |>
  make_backwards_elimination_table('y', c('x1', 'x2', 'x3', 'x4'))
}
