% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_p_value_table.R
\name{make_p_value_table}
\alias{make_p_value_table}
\title{Make p-value Table}
\usage{
make_p_value_table(
  tibble_to_analyze,
  dependent_variable,
  independent_variables,
  p_value_column_name = "p_value",
  initial_p_value_column_name = "Pr(>|t|)",
  term_column_name = "term"
)
}
\arguments{
\item{tibble_to_analyze}{Tibble containing dependent and independent variables.}

\item{dependent_variable}{Dependent variable in \code{tibble_to_analyze}.}

\item{independent_variables}{List of independent variables.}

\item{p_value_column_name}{Name of the output p-value column.}

\item{initial_p_value_column_name}{Name of column of p-values output when running \code{summary()} on a generalized linear model.}

\item{term_column_name}{Name of column of terms output when running \code{summary()} on a generalized linear model.}
}
\value{
A tibble of terms and p-values.
}
\description{
Make a table of p-value for terms of a generalized linear model.
}
\examples{
tibble::tibble(a = c(1, 2, 3), b = c(1, 2, 3), c = c(1, 2, 3)) |>
  make_p_value_table('c', c('a', 'b'))
}
