% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{performance_metrics}
\alias{performance_metrics}
\title{Performance Metrics}
\usage{
performance_metrics(
  total_population = integer(),
  real_positives = integer(),
  real_negatives = integer(),
  true_positives = integer(),
  false_positives = integer(),
  true_negatives = integer(),
  false_negatives = integer()
)
}
\arguments{
\item{total_population}{The number of people in the entire population.}

\item{real_positives}{The number of people with the condition in the population.}

\item{real_negatives}{The number of people without the condition in the population.}

\item{true_positives}{The number of people with the condition with a positive test result.}

\item{false_positives}{The number of people without the condition with a positive test result.}

\item{true_negatives}{The number of people without the condition with a negative test result.}

\item{false_negatives}{The number of people without the condition with a positive test result.}
}
\value{
A list of arguments and calculated performance metrics (see table).
}
\description{
Takes a combination of values typically found in a confusion matrix and calculates the predictive performance metrics described on \href{https://en.wikipedia.org/wiki/Confusion_matrix#Table_of_confusion}{this Wikipedia page}.
}
\details{
The function is meant as a quick way to get multiple performance metrics while experimenting with models. Nomenclature of return values is therefore informal with several redundant names for certain values, letting the the user choose their preferred terminology.

Arguments are returned together with calculated values. Missing arguments are calculated, as far as sufficient data is provided. The function throws an error if insufficient data or contradictory data is provided.
}
\examples{
# The function can be called without all arguments, as long as sufficient data is provided.
performance_metrics(
  total_population = 2030,
  real_positives = 30,
  real_negatives = 2000,
  true_positives = 20,
  false_negatives = 10,
  false_positives = 180,
  true_negatives = 1820
)

performance_metrics(
  total_population = 2030,
  real_positives = 30,
  real_negatives = 2000,
  true_positives = 20,
  false_negatives = 10,
  false_positives = 180
)

performance_metrics(
  total_population = 2030,
  real_positives = 30,
  real_negatives = 2000,
  true_positives = 20,
  false_positives = 180
)

performance_metrics(
  real_positives = 30,
  real_negatives = 2000,
  true_positives = 20,
  false_positives = 180
)
}
