% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_highest_p_value.R
\name{remove_highest_p_value}
\alias{remove_highest_p_value}
\title{Remove Highest p-value}
\usage{
remove_highest_p_value(
  p_value_table,
  independent_variables_column_name = "independent_variables"
)
}
\arguments{
\item{p_value_table}{A p-value table returned by \code{make_p_value_table()}.}

\item{independent_variables_column_name}{List of independent variables.}
}
\value{
A list of term names.
}
\description{
Take a p-value table returned by \code{make_p_value_table()} and return a list of terms excluding the one with the highest value, excluding the intercept.
}
\examples{
tibble::tibble(
  independent_variables = c('(Intercept)', 'x1', 'x2', 'x3', 'x4'),
  p_value = c(1.18e-5, 4.36e-33, 4.97e-14, 6.02e-23, 8.44e-1)
) |>
  remove_highest_p_value()
}
\seealso{
\code{\link[=make_p_value_table]{make_p_value_table()}}
}
